move code handling to 'io'
